Moved latest summary state tracking from garbage collector into a separate class
